Add `extend` method
This allows users to add, customize and override tests in their
projects. Users could extend the library to add their own tests, such as
if Cookies are enabled for the browser, for example (#46). Users could
change the way a test approaches its detection measures to be more
strict or accurate. Finally a user could fix a bug in their project as
a stop-gap measure between releases of the library.